+++ experiments/split_cifar10/er_aml.py
@@ -5,16 +5,12 @@ import torchvision.transforms as transforms
 from torch.optim import SGD
 
 from avalanche.benchmarks.classic import SplitCIFAR10
-from avalanche.benchmarks.generators import benchmark_with_validation_stream
-from avalanche.benchmarks.scenarios.online_scenario import OnlineCLScenario
 from avalanche.evaluation.metrics import accuracy_metrics, loss_metrics
 from avalanche.logging import InteractiveLogger
-from avalanche.models import SlimResNet18
-from avalanche.models.dynamic_modules import IncrementalClassifier
 from avalanche.training.plugins import EvaluationPlugin
-from avalanche.training.supervised import OnlineNaive, ER_AML
+from avalanche.training.supervised import ER_AML
 from experiments.utils import create_default_args, set_seed
-from models.reduced_resnet18 import SingleHeadReducedResNet18
+from models import SingleHeadReducedResNet18
 
 
 def eraml_scifar10(override_args=None):
