+++ tests/target_results.csv
@@ -32,5 +32,5 @@ online_replay, scifar100, 0.21
 online_replay, scifar10, 0.50
 online_replay, smnist, 0.92
 scr,scifar10,0.36
-er_aml,scifar10,0.52
+er_aml,scifar10,0.47
 er_aml,scifar100,0.24
